Conflicts:
	NewBasicTemplate2016/src/CommandBase.cpp
	NewBasicTemplate2016/src/CommandBase.h
	NewBasicTemplate2016/src/Commands/RunningOverDefense.cpp
	NewBasicTemplate2016/src/Commands/UltrasonicTesting.cpp
	NewBasicTemplate2016/src/Commands/UltrasonicTesting.h
	NewBasicTemplate2016/src/OI.cpp
	NewBasicTemplate2016/src/Robot.cpp
	NewBasicTemplate2016/src/RobotMap.h
	NewBasicTemplate2016/src/Subsystems/AnalogSensors.cpp
	NewBasicTemplate2016/src/Subsystems/AnalogSensors.h
	NewBasicTemplate2016/src/Subsystems/DriveTrain.cpp
	NewBasicTemplate2016/src/Subsystems/DriveTrain.h
	NewBasicTemplate2016/src/Subsystems/SerialCommunication.cpp
	NewBasicTemplate2016/sysProps.xml

+++ NewBasicTemplate2016/src/Commands/UltrasonicTesting.h
@@ -7,12 +7,11 @@
 class UltrasonicTesting: public CommandBase
 {
 private:
-	double leftSonarDistance,
-		   rightSonarDistance;
+	float leftSonarDistance,
+		  rightSonarDistance;
 
 	int lightSensorDistance;
 
-	//  int i;
 public:
 	UltrasonicTesting();
 	void Initialize();
